refactor(gemini): move Gemini Vertex AI classes into dedicated package
