+++ src/main/java/org/phong/zenflow/plugin/subdomain/nodes/builtin/integration/google/ai/gemini/vertex/schema.vertex.json
